@@ -132,7 +123,7 @@
     "supertest": "^3.0.0",
     "ts-node": "^6.0.0",
     "tslint": "^5.11.0",
-    "typescript": "^3.1.6"
+    "typescript": "^3.2.2"
   },
   "collective": {
     "type": "opencollective",
