@@ -1,8 +1,8 @@
 import { NestInterceptor } from '@nestjs/common';
-import { Controller } from '@nestjs/common/interfaces';
+import { CallHandler, Controller } from '@nestjs/common/interfaces';
 import { isEmpty } from '@nestjs/common/utils/shared.utils';
 import { defer, from as fromPromise, Observable } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
+import { mergeAll, switchMap } from 'rxjs/operators';
 import { ExecutionContextHost } from '../helpers/execution-context.host';
 
 export class InterceptorsConsumer {
@@ -18,20 +18,17 @@ export class InterceptorsConsumer {
     }
     const context = this.createContext(args, instance, callback);
     const start$ = defer(() => this.transformDeffered(next));
-    /***
-      const nextFn =  (i: number) => async () => {
-      if (i <= interceptors.length) {
+
+    const nextFn = (i = 0) => async () => {
+      if (i >= interceptors.length) {
         return start$;
       }
-      return await interceptors[i].intercept(context, nextFn(i + 1) as any);
+      const handler: CallHandler = {
+        handle: () => fromPromise(nextFn(i + 1)()).pipe(mergeAll()),
+      };
+      return interceptors[i].intercept(context, handler);
     };
-    */
-    const result$ = await interceptors.reduce(
-      async (stream$, interceptor) =>
-        interceptor.intercept(context, await stream$),
-      Promise.resolve(start$),
-    );
-    return result$.toPromise();
+    return nextFn()();
   }
 
   public createContext(
