@@ -15,17 +15,17 @@ export function Inject<T = any>(token?: T) {
       token && isFunction(token) ? ((token as any) as Function).name : token;
 
     if (!isUndefined(index)) {
-      const dependencies =
+      let dependencies =
         Reflect.getMetadata(SELF_DECLARED_DEPS_METADATA, target) || [];
 
-      dependencies.push({ index, param: type });
+      dependencies = [...dependencies, { index, param: type }];
       Reflect.defineMetadata(SELF_DECLARED_DEPS_METADATA, dependencies, target);
       return;
     }
-    const properties =
+    let properties =
       Reflect.getMetadata(PROPERTY_DEPS_METADATA, target.constructor) || [];
 
-    properties.push({ key, type });
+    properties = [...properties, { key, type }];
     Reflect.defineMetadata(
       PROPERTY_DEPS_METADATA,
       properties,
