@@ -1,9 +1,13 @@
 import { Observable } from 'rxjs';
 import { ExecutionContext } from './execution-context.interface';
 
+export interface CallHandler<T = any> {
+  handle(): Observable<T>;
+}
+
 export interface NestInterceptor<T = any, R = any> {
   intercept(
     context: ExecutionContext,
-    call$: Observable<T>,
+    next: CallHandler<T>,
   ): Observable<R> | Promise<Observable<R>>;
 }
