@@ -131,7 +131,7 @@ export class NestApplication extends NestApplicationContext
     await this.registerMiddleware(this.httpAdapter);
 
     const prefix = this.config.getGlobalPrefix();
-    const basePath = prefix ? validatePath(prefix) : '';
+    const basePath = validatePath(prefix);
     this.routesResolver.resolve(this.httpAdapter, basePath);
   }
 
