@@ -1,5 +1,7 @@
+import { isFunction } from '@nestjs/common/utils/shared.utils';
 import { ExecutionContextHost } from '@nestjs/core/helpers/execution-context.host';
 import { Observable } from 'rxjs';
+import { catchError } from 'rxjs/operators';
 import { RpcExceptionsHandler } from '../exceptions/rpc-exceptions-handler';
 
 export class RpcProxy {
@@ -9,11 +11,30 @@ export class RpcProxy {
   ): (...args: any[]) => Promise<Observable<any>> {
     return async (...args: any[]) => {
       try {
-        return await targetCallback(...args);
-      } catch (e) {
-        const host = new ExecutionContextHost(args);
-        return exceptionsHandler.handle(e, host);
+        const result = await targetCallback(...args);
+        return !this.isObservable(result)
+          ? result
+          : result.pipe(
+              catchError(error =>
+                this.handleError(exceptionsHandler, args, error),
+              ),
+            );
+      } catch (error) {
+        return this.handleError(exceptionsHandler, args, error);
       }
     };
+  }
+
+  handleError<T>(
+    exceptionsHandler: RpcExceptionsHandler,
+    args: any[],
+    error: T,
+  ): Observable<any> {
+    const host = new ExecutionContextHost(args);
+    return exceptionsHandler.handle(error, host);
+  }
+
+  isObservable(result: any): boolean {
+    return result && isFunction(result.subscribe);
   }
 }
