@@ -1,8 +1,8 @@
-export * from './client-metadata.interface';
-export * from './microservice-configuration.interface';
-export * from './pattern-metadata.interface';
-export * from './custom-transport-strategy.interface';
-export * from './message-handlers.interface';
-export * from './closeable.interface';
-export * from './packet.interface';
 export * from './client-grpc.interface';
+export * from './client-metadata.interface';
+export * from './closeable.interface';
+export * from './custom-transport-strategy.interface';
+export * from './message-handler.interface';
+export * from './microservice-configuration.interface';
+export * from './packet.interface';
+export * from './pattern-metadata.interface';
