@@ -1,6 +1,10 @@
 import { fromEvent } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
-import { CANCEL_EVENT, GRPC_DEFAULT_URL } from '../constants';
+import {
+  CANCEL_EVENT,
+  GRPC_DEFAULT_PROTO_LOADER,
+  GRPC_DEFAULT_URL,
+} from '../constants';
 import { InvalidGrpcPackageException } from '../errors/invalid-grpc-package.exception';
 import { InvalidProtoDefinitionException } from '../errors/invalid-proto-definition.exception';
 import { CustomTransportStrategy } from '../interfaces';
@@ -28,11 +32,12 @@ export class ServerGrpc extends Server implements CustomTransportStrategy {
     this.url =
       this.getOptionsProp<GrpcOptions>(options, 'url') || GRPC_DEFAULT_URL;
 
+    const protoLoader =
+      this.getOptionsProp<GrpcOptions>(options, 'protoLoader') ||
+      GRPC_DEFAULT_PROTO_LOADER;
+
     grpcPackage = this.loadPackage('grpc', ServerGrpc.name);
-    grpcProtoLoaderPackage = this.loadPackage(
-      '@grpc/proto-loader',
-      ServerGrpc.name,
-    );
+    grpcProtoLoaderPackage = this.loadPackage(protoLoader, ServerGrpc.name);
   }
 
   public async listen(callback: () => void) {
@@ -75,9 +80,9 @@ export class ServerGrpc extends Server implements CustomTransportStrategy {
 
     // tslint:disable-next-line:forin
     for (const methodName in grpcService.prototype) {
-      const methodHandler = this.messageHandlers[
-        this.createPattern(name, methodName)
-      ];
+      const methodHandler = this.getHandlerByPattern(
+        this.createPattern(name, methodName),
+      );
       if (!methodHandler) {
         continue;
       }
