@@ -1,9 +1,9 @@
-import * as sinon from 'sinon';
 import { expect } from 'chai';
+import { of, throwError } from 'rxjs';
+import * as sinon from 'sinon';
 import { RpcProxy } from '../../context/rpc-proxy';
-import { RpcExceptionsHandler } from '../../exceptions/rpc-exceptions-handler';
 import { RpcException } from '../../exceptions/rpc-exception';
-import { Observable, of } from 'rxjs';
+import { RpcExceptionsHandler } from '../../exceptions/rpc-exceptions-handler';
 
 describe('RpcProxy', () => {
   let routerProxy: RpcProxy;
@@ -18,10 +18,7 @@ describe('RpcProxy', () => {
 
   describe('create', () => {
     it('should method return thunk', async () => {
-      const proxy = await routerProxy.create(
-        async data => of(true),
-        handler,
-      );
+      const proxy = await routerProxy.create(async data => of(true), handler);
       expect(typeof proxy === 'function').to.be.true;
     });
 
@@ -32,6 +29,27 @@ describe('RpcProxy', () => {
       }, handler);
       await proxy(null);
       expectation.verify();
+    });
+
+    it('should attach "catchError" operator when observable was returned', async () => {
+      const expectation = handlerMock.expects('handle').once();
+      const proxy = routerProxy.create(async (client, data) => {
+        return throwError(new RpcException('test'));
+      }, handler);
+      (await proxy(null, null)).subscribe(null, () => expectation.verify());
+    });
+  });
+
+  describe('isObservable', () => {
+    describe('when observable', () => {
+      it('should return true', () => {
+        expect(routerProxy.isObservable(of('test'))).to.be.true;
+      });
+    });
+    describe('when not observable', () => {
+      it('should return false', () => {
+        expect(routerProxy.isObservable({})).to.be.false;
+      });
     });
   });
 });
