@@ -1,4 +1,5 @@
 import {
+  CallHandler,
   ExecutionContext,
   Inject,
   mixin,
@@ -36,7 +37,7 @@ export function FilesInterceptor(
 
     async intercept(
       context: ExecutionContext,
-      call$: Observable<any>,
+      next: CallHandler,
     ): Promise<Observable<any>> {
       const ctx = context.switchToHttp();
 
@@ -53,7 +54,7 @@ export function FilesInterceptor(
           },
         ),
       );
-      return call$;
+      return next.handle();
     }
   }
   const Interceptor = mixin(MixinInterceptor);
