@@ -1,4 +1,9 @@
-import { Injectable, NestInterceptor, ExecutionContext } from '@nestjs/common';
+import {
+  CallHandler,
+  ExecutionContext,
+  Injectable,
+  NestInterceptor,
+} from '@nestjs/common';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
@@ -11,8 +16,8 @@ export class TransformInterceptor<T>
   implements NestInterceptor<T, Response<T>> {
   intercept(
     context: ExecutionContext,
-    call$: Observable<T>,
+    next: CallHandler<T>,
   ): Observable<Response<T>> {
-    return call$.pipe(map(data => ({ data })));
+    return next.handle().pipe(map(data => ({ data })));
   }
 }
