@@ -1,28 +1,28 @@
 import {
+  CallHandler,
+  ExecutionContext,
+  HttpException,
+  HttpStatus,
   Injectable,
   NestInterceptor,
-  ExecutionContext,
-  HttpStatus,
 } from '@nestjs/common';
-import { HttpException } from '@nestjs/common';
 import { Observable, throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 
 @Injectable()
 export class ExceptionInterceptor implements NestInterceptor {
-  intercept(
-    context: ExecutionContext,
-    call$: Observable<any>,
-  ): Observable<any> {
-    return call$.pipe(
-      catchError(err =>
-        throwError(
-          new HttpException(
-            'Exception interceptor message',
-            HttpStatus.BAD_GATEWAY,
+  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
+    return next
+      .handle()
+      .pipe(
+        catchError(err =>
+          throwError(
+            new HttpException(
+              'Exception interceptor message',
+              HttpStatus.BAD_GATEWAY,
+            ),
           ),
         ),
-      ),
-    );
+      );
   }
 }
